requirements.txt
asgiref==3.5.0
Django==4.0.4
django-material==1.10.0
django-ranged-response==0.2.0
django-session-timeout==0.1.0
django-simple-captcha==0.5.17
mysqlclient==2.1.0
Pillow==9.1.0
pyotp==2.6.0
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
